// BetacomareFront/src/app/features/models/product.ts
export interface Product {
    productId:              number;
    name:                   string;
    productNumber:          string;
    color:                  string;
    standardCost:           number;
    listPrice:              number;
    size:                   string;
    weight:                 number;
    productCategoryId:      number;
    productModelId:         number;
    sellStartDate:          string;
    sellEndDate:            null;
    discontinuedDate:       null;
    thumbNailPhoto:         string;
    thumbnailPhotoFileName: string;
    rowguid:                string;
    modifiedDate:           string;
    productCategory:        null;
    productModel:           null;
    salesOrderDetails:      SalesOrderDetails;
}

export interface SalesOrderDetails {
    $values: any[];
}